1.2.2

+++ src/resources/spaces/spaces.resource.ts
@@ -1,6 +1,8 @@
+import { hasRolePermission, hasSpace } from "@database/functions/space";
+import { UserModelType, UserSpaceType } from "@utils/types/models/user";
+import { hasExistsUser, hasUser } from "@database/functions/user";
 import { ManageRequestBody } from "@middlewares/manageRequest";
 import stringService from "@utils/services/stringServices";
-import { hasSpace } from "@database/functions/space";
 import spaceModel from "@database/model/space";
 import userModel from "@database/model/user";
 
@@ -41,7 +43,7 @@ const spacesResource = {
             manageError({ code: "internal_error", error });
         }
     },
-    createSpaceRole: async ({ manageError, params, data }: ManageRequestBody) => {
+    createSpaceRole: async ({ manageError, params, data, ids }: ManageRequestBody) => {
         try {
             const { spaceID } =  params;
             if (!spaceID) return manageError({ code: "invalid_params" });
@@ -51,6 +53,15 @@ const spacesResource = {
 
             let { name, permissions } = data;
             if (!name || !permissions) return manageError({ code: "invalid_data" });
+
+            const { userID } = ids;
+
+            const user = await hasUser({ _id: userID }, manageError);
+            if (!user) return;
+
+            const userSpace = user.spaces?.find(x => x.id == spaceID);
+            const hasPermisson = await hasRolePermission(userSpace?.role.toString() || "", ["administrator", "manage_roles", "owner"]);
+            if (!hasPermisson) return manageError({ code: "no_execution_permission" });
 
             name = stringService.removeSpacesAndLowerCase(name);
 
@@ -71,14 +82,173 @@ const spacesResource = {
             manageError({ code: "internal_error", error });
         }
     },
-    updateSpaceRole: async ({ manageError, params, data }: ManageRequestBody) => {
+    createSpaceUser: async ({ manageError, params, data, ids }: ManageRequestBody) => {
         try {
-            const { spaceID, roleID } =  params;
-            if (!spaceID || !roleID) return manageError({ code: "invalid_params" });
+            const { spaceID } =  params;
+            if (!spaceID) return manageError({ code: "invalid_params" });
 
             const space = await hasSpace({ _id: spaceID }, manageError);
             if (!space) return;
 
+            let { id, name, roleID } = data;
+            if (!id || !name) return manageError({ code: "invalid_data" });
+
+            const { userID } = ids;
+
+            const user = await hasUser({ _id: userID }, manageError);
+            if (!user) return;
+
+            const userSpace = user.spaces?.find(x => x.id == spaceID);
+            const hasPermisson = await hasRolePermission(userSpace?.role.toString() || "", ["administrator", "manage_users", "owner"]);
+            if (!hasPermisson) return manageError({ code: "no_execution_permission" });
+
+
+            id = stringService.removeSpacesAndLowerCase(id);
+            name = stringService.normalizeString(name);
+
+            const userExists = await hasExistsUser({ id }, manageError);
+            if (!userExists) return;
+
+            const extra: Partial<UserModelType> = {
+                lastUpdate: new Date(Date.now()),
+            };
+
+            const normalRole = Array.isArray(space.roles) ? space.roles.find((x) =>  x.name == "normal") : null;
+
+            let newSpace: UserSpaceType= {
+                entryAt: new Date(),
+                role: roleID ? roleID : normalRole._id,
+                name: space.name,
+                id: space._id
+            };
+        
+            extra.spaces = [newSpace];
+
+            let spaceUserMetrics = space.metrics?.users || 0;
+
+            await spaceModel.findByIdAndUpdate(space._id, { $set:{ metrics: { user: spaceUserMetrics + 1 } } }, { new: true });
+            
+            const createdUser = new userModel({ id, name, ...extra });
+            await createdUser.save();
+
+            return {
+                user: createdUser
+            };
+        } catch (error) {
+            manageError({ code: "internal_error", error });
+        }
+    },
+    addSpaceUser: async ({ manageError, params, data, ids }: ManageRequestBody) => {
+        try {
+            const { spaceID } =  params;
+            if (!spaceID) return manageError({ code: "invalid_params" });
+
+            const space = await hasSpace({ _id: spaceID }, manageError);
+            if (!space) return;
+
+            let { id, roleID } = data;
+            if (!id) return manageError({ code: "invalid_data" });
+
+            const { userID } = ids;
+
+            const user = await hasUser({ _id: userID }, manageError);
+            if (!user) return;
+
+            const userSpace = user.spaces?.find(x => x.id == spaceID);
+            const hasPermisson = await hasRolePermission(userSpace?.role.toString() || "", ["administrator", "manage_users", "owner"]);
+            if (!hasPermisson) return manageError({ code: "no_execution_permission" });
+
+
+            id = stringService.removeSpacesAndLowerCase(id);
+
+            const invitedUser = await hasUser({ id }, manageError);
+            if (!invitedUser) return;
+
+            const extra: Partial<UserModelType> = {
+                lastUpdate: new Date(Date.now()),
+            };
+
+            const normalRole = Array.isArray(space.roles) ? space.roles.find((x) =>  x.name == "normal") : null;
+
+            let newSpace: UserSpaceType= {
+                role: roleID ? roleID : normalRole._id,
+                entryAt: new Date(),
+                name: space.name,
+                id: space._id
+            };
+        
+            extra.spaces = [...(invitedUser.spaces || []), newSpace];
+
+            let spaceUserMetrics = space.metrics?.users || 0;
+
+            await spaceModel.findByIdAndUpdate(space._id, { $set:{ metrics: { user: spaceUserMetrics + 1 } } }, { new: true });
+            
+            const updatedUser  = await userModel.findByIdAndUpdate(invitedUser._id, { $set:{ ...invitedUser, extra } }, { new: true }).select("-password");
+
+            return {
+                user: updatedUser
+            };
+        } catch (error) {
+            manageError({ code: "internal_error", error });
+        }
+    },
+    removeSpaceUser: async ({ manageError, params, data, ids }: ManageRequestBody) => {
+        try {
+            const { spaceID } =  params;
+            if (!spaceID) return manageError({ code: "invalid_params" });
+
+            const space = await hasSpace({ _id: spaceID }, manageError);
+            if (!space) return;
+
+            let { id } = data;
+            if (!id) return manageError({ code: "invalid_data" });
+
+            const { userID } = ids;
+
+            const user = await hasUser({ _id: userID }, manageError);
+            if (!user) return;
+
+            const userSpace = user.spaces?.find(x => x.id == spaceID);
+            const hasPermisson = await hasRolePermission(userSpace?.role.toString() || "", ["administrator", "manage_users", "owner"]);
+            if (!hasPermisson) return manageError({ code: "no_execution_permission" });
+
+            id = stringService.removeSpacesAndLowerCase(id);
+
+            const removeUser = await hasUser({ id }, manageError);
+            if (!removeUser) return;
+
+            const spaces = removeUser.spaces?.filter(x => x.id == spaceID);
+        
+            let spaceUserMetrics = space.metrics?.users || 0;
+
+            await spaceModel.findByIdAndUpdate(space._id, { $set:{ metrics: { user: spaceUserMetrics - 1 } } }, { new: true });
+            
+            const removedUser  = await userModel.findByIdAndUpdate(removeUser._id, { $set:{ spaces, lastUpdate: Date.now() } }, { new: true }).select("-password");
+
+            return {
+                user: removedUser
+            };
+        } catch (error) {
+            manageError({ code: "internal_error", error });
+        }
+    },
+    updateSpaceRole: async ({ manageError, params, data, ids }: ManageRequestBody) => {
+        try {
+            const { spaceID, roleID } =  params;
+            const { userID } =  ids;
+            if (!spaceID || !roleID || !userID) return manageError({ code: "invalid_params" });
+
+            const user = await hasUser({ _id: userID }, manageError);
+            if (!user) return;
+            
+            const space = await hasSpace({ _id: spaceID }, manageError);
+
+            if (!space) return;
+
+            const userSpace = user.spaces?.find(x => x.id == spaceID);
+            const hasPermisson = await hasRolePermission(userSpace?.role.toString() || "", ["administrator", "manage_roles", "owner"]);
+            if (!hasPermisson) return manageError({ code: "no_execution_permission" });
+ 
             const spaceRole = Array.isArray(space.roles) ? space.roles.find((x) => String(x._id) === roleID) : null;
             if (!spaceRole) return manageError({ code: "role_not_found" });
 
@@ -119,13 +289,22 @@ const spacesResource = {
             manageError({ code: "internal_error", error });
         }
     },
-    deleteSpaceRole: async ({ manageError, params }: ManageRequestBody) => {
+    deleteSpaceRole: async ({ manageError, params, ids }: ManageRequestBody) => {
         try {
             const { spaceID, roleID } =  params;
             if (!spaceID || !roleID) return manageError({ code: "invalid_params" });
 
             const space = await hasSpace({ _id: spaceID }, manageError);
             if (!space) return;
+
+            const { userID } = ids;
+
+            const user = await hasUser({ _id: userID }, manageError);
+            if (!user) return;
+
+            const userSpace = user.spaces?.find(x => x.id == spaceID);
+            const hasPermisson = await hasRolePermission(userSpace?.role.toString() || "", ["administrator", "manage_roles", "owner"]);
+            if (!hasPermisson) return manageError({ code: "no_execution_permission" });
 
             const spaceRole = Array.isArray(space.roles) ? space.roles.find((x) => String(x._id) === roleID) : null;
             if (!spaceRole) return manageError({ code: "role_not_found" });
@@ -150,7 +329,7 @@ const spacesResource = {
             manageError({ code: "internal_error", error });
         }
     },
-    configSpaceModule: async ({ manageError, params, data }: ManageRequestBody) => {
+    configSpaceModule: async ({ manageError, params, data, ids }: ManageRequestBody) => {
         try {
             const { spaceID, module } = params;
             if (!spaceID || !module) return manageError({ code: "invalid_params" });
@@ -164,6 +343,15 @@ const spacesResource = {
     
             const currentModule = space.modules[module as keyof typeof space.modules] as any;
             if (!currentModule) return manageError({ code: "invalid_params" });
+
+            const { userID } = ids;
+
+            const user = await hasUser({ _id: userID }, manageError);
+            if (!user) return;
+
+            const userSpace = user.spaces?.find(x => x.id == spaceID);
+            const hasPermisson = await hasRolePermission(userSpace?.role.toString() || "", ["administrator", "manage_modules", "owner"]);
+            if (!hasPermisson) return manageError({ code: "no_execution_permission" });
 
             const { config } = data;
